[feat] navigation완성
