+++ app/src/main/java/com/hmh/hamyeonham/presentation/main/MainActivity.kt
@@ -2,7 +2,7 @@ package com.hmh.hamyeonham.presentation.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.fragment.findNavController
+import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.hmh.hamyeonham.R
@@ -16,13 +16,18 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun initNavigation() {
-        val navController =
-            supportFragmentManager.findFragmentById(
-                androidx.navigation.fragment.R.id.nav_host_fragment_container,
-            )?.findNavController()
-        navController?.let {
-            val bottomNav = findViewById<BottomNavigationView>(R.id.btn_main)
-            bottomNav.setupWithNavController(navController)
-        }
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        val bottomNavView = findViewById<BottomNavigationView>(R.id.btn_main)
+        bottomNavView.setupWithNavController(navController)
+//        val navController =
+//            supportFragmentManager.findFragmentById(
+//                androidx.navigation.fragment.R.id.nav_host_fragment_container,
+//            )?.findNavController()
+//        navController?.let {
+//            val bottomNav = findViewById<BottomNavigationView>(R.id.btn_main)
+//            bottomNav.setupWithNavController(navController)
+//        }
     }
 }
