+++ app/src/main/java/com/hmh/hamyeonham/presentation/main/UsagestatFragment.kt
@@ -1,0 +1,19 @@
+package com.hmh.hamyeonham.presentation.main
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.hmh.hamyeonham.R
+
+class UsagestatFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_usagestat, container, false)
+    }
+}
