--- app/src/main/java/com/hmh/hamyeonham/presentation/usagestatistics/UsagestatisticsFragment.kt
@@ -1,57 +1,0 @@
-package com.hmh.hamyeonham.presentation.usagestatistics
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.hmh.hamyeonham.databinding.FragmentUsagestatisticsBinding
-
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-class UsagestatisticsFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-//        arguments?.let {
-//            param1 = it.getString(ARG_PARAM1)
-//            param2 = it.getString(ARG_PARAM2)
-//        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? {
-        // Inflate the layout for this fragment
-        val binding = FragmentUsagestatisticsBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-//    companion object {
-//        /**
-//         * Use this factory method to create a new instance of
-//         * this fragment using the provided parameters.
-//         *
-//         * @param param1 Parameter 1.
-//         * @param param2 Parameter 2.
-//         * @return A new instance of fragment MypageFragment.
-//         */
-//        @JvmStatic
-//        fun newInstance(
-//            param1: String,
-//            param2: String,
-//        ) = MypageFragment().apply {
-//            arguments =
-//                Bundle().apply {
-//                    putString(ARG_PARAM1, param1)
-//                    putString(ARG_PARAM2, param2)
-//                }
-//        }
-//    }
-}
