+++ app/src/main/res/menu/menu_main.xml
@@ -1,15 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/usagestatisticsFragment"
+        android:id="@+id/achievFragment"
         android:icon="@drawable/ic_launcher_foreground"
-        android:title="usagestatistics"/>
+        android:title="달성현황" />
+    <item
+        android:id="@+id/usagestatFragment"
+        android:icon="@drawable/ic_launcher_foreground"
+        android:title="이용시간" />
     <item
         android:id="@+id/mypageFragment"
         android:icon="@drawable/ic_launcher_foreground"
-        android:title="mypage"/>
-    <item
-        android:id="@+id/achievmentFragment"
-        android:icon="@drawable/ic_launcher_foreground"
-        android:title="achievment"/>
+        android:title="my page" />
+
 </menu>
