--- app/src/main/res/navigation/nav_graph.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_graph">
-
-    <fragment
-        android:id="@+id/usagestatisticsFragment"
-        android:name="com.hmh.hamyeonham.presentation.usagestatistics.UsagestatisticsFragment"
-        android:label="fragment_usagestatistics"
-        tools:layout="@layout/fragment_usagestatistics">
-    </fragment>
-    <fragment
-        android:id="@+id/mypageFragment"
-        android:name="com.hmh.hamyeonham.presentation.mypage.MypageFragment"
-        android:label="fragment_mypage"
-        tools:layout="@layout/fragment_mypage">
-
-    </fragment>
-    <fragment
-        android:id="@+id/achievmentFragment"
-        android:name="com.hmh.hamyeonham.presentation.achievment.AchievmentFragment"
-        android:label="fragment_achievment"
-        tools:layout="@layout/fragment_achievment">
-    </fragment>
-</navigation>
